restructured utility components and removed unused components
