+++ src/components/interfaces/ask-noora/Result.tsx
