+++ src/components/interfaces/noora-chat/menu/sections/Technical.tsx
