+++ src/components/interfaces/noora-chat/summary/Footer.tsx
