--- src/pages/rate-limiting-test.tsx
@@ -1,6 +1,0 @@
-import React from "react";
-import RateLimitingTest from "../components/global/utility/RateLimitingTest";
-
-export default function ratelimitingtest() {
-  return <RateLimitingTest />;
-}
